requirements.txt
Django~=4.2.7
django-jazzmin~=2.6.0
django-q==1.3.9
psycopg2-binary==2.9.9
whitenoise==6.6.0
PyJWT==2.8.0
